requirements.txt
flask
onnxruntime
requests
numpy
Pillow
